requirements.txt
keyboard==0.13.5
numpy==2.2.1
opencv_python==4.10.0.84
Pillow==11.1.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
pynput==1.7.7
pyperclip==1.9.0
PyYAML==6.0.2
